fix(app): keep controller compose lintable
