+++ packages/app/src/docker-git/controller-compose.ts
@@ -4,26 +4,31 @@ import * as FileSystem from "@effect/platform/FileSystem"
 import * as Path from "@effect/platform/Path"
 import { Duration, Effect } from "effect"
 
+import {
+  type ControllerComposeFiles,
+  type ControllerGpuMode,
+  composeFilesForGpuMode,
+  controllerGpuModeEnvKey,
+  loadControllerComposeExtraPath
+} from "./controller-compose-files.js"
 import { loadControllerDockerRuntime, resolveControllerRuntimeOverlayPath } from "./controller-compose-runtime.js"
 import { computeLocalControllerRevision, controllerRevisionEnvKey } from "./controller-revision.js"
 import type { ControllerDockerRuntime } from "./controller-runtime.js"
 import { runCommandWithCapturedOutput } from "./frontend-lib/shell/command-runner.js"
 import { findExistingUpwards } from "./frontend-lib/usecases/path-helpers.js"
-import type { ControllerBootstrapError } from "./host-errors.js"
+import { type ControllerBootstrapError, controllerBootstrapError } from "./host-errors.js"
 
-export const controllerGpuModeEnvKey = "DOCKER_GIT_CONTROLLER_GPU"
 export const controllerBuildSkillerEnvKey = "DOCKER_GIT_CONTROLLER_BUILD_SKILLER"
-export const controllerComposeExtraFileEnvKey = "DOCKER_GIT_CONTROLLER_COMPOSE_EXTRA_FILE"
 
-export type ControllerGpuMode = "none" | "all"
 export type ControllerBuildSkillerMode = "0" | "1"
 
-export type ControllerComposeFiles = {
-  readonly composePath: string
-  readonly extraOverlayPath: string | null
-  readonly gpuOverlayPath: string | null
-  readonly runtimeOverlayPath: string | null
-}
+export {
+  composeFilesForMode,
+  composeFilesToArgs,
+  controllerComposeExtraFileEnvKey,
+  controllerGpuModeEnvKey
+} from "./controller-compose-files.js"
+export type { ControllerComposeFiles, ControllerGpuMode } from "./controller-compose-files.js"
 
 export const controllerComposeProjectName = "docker-git"
 
@@ -44,11 +49,6 @@ export const controllerComposeProjectArgs: ReadonlyArray<string> = [
 
 const skillerSubmodulePath = "third_party/skiller-desktop-skills-manager"
 const skillerPackagePath = `${skillerSubmodulePath}/package.json`
-
-const controllerBootstrapError = (message: string): ControllerBootstrapError => ({
-  _tag: "ControllerBootstrapError",
-  message
-})
 
 export const parseControllerGpuMode = (raw?: string): ControllerGpuMode | null => {
   const trimmed = raw?.trim() ?? ""
@@ -115,42 +115,6 @@ const mapSkillerPathError = (error: PlatformError): ControllerBootstrapError =>
 
 const mapControllerRevisionError = (error: PlatformError): ControllerBootstrapError =>
   controllerBootstrapError(`Failed to compute docker-git controller revision.\nDetails: ${String(error)}`)
-
-// CHANGE: add a verified controller compose overlay boundary for E2E/runtime callers
-// WHY: temporary compose overrides must be part of the explicit docker compose argument vector
-// QUOTE(ТЗ): n/a
-// REF: issue-440-review-compose-overlay
-// SOURCE: n/a
-// FORMAT THEOREM: forall p: env(extra)=p and exists(resolve(p)) -> resolve(extra)=Some(resolve(p))
-// PURITY: SHELL
-// EFFECT: Effect<string | null, ControllerBootstrapError, FileSystem | Path>
-// INVARIANT: non-empty extra compose env values either resolve to an existing file or fail before docker compose
-// COMPLEXITY: O(1)
-const loadControllerComposeExtraPath = (): Effect.Effect<
-  string | null,
-  ControllerBootstrapError,
-  FileSystem.FileSystem | Path.Path
-> =>
-  Effect.gen(function*(_) {
-    const raw = process.env[controllerComposeExtraFileEnvKey]?.trim() ?? ""
-    if (raw.length === 0) {
-      return null
-    }
-
-    const fs = yield* _(FileSystem.FileSystem)
-    const path = yield* _(Path.Path)
-    const extraOverlayPath = path.resolve(raw)
-    const isExists = yield* _(fs.exists(extraOverlayPath).pipe(Effect.mapError(mapComposePathError)))
-    return isExists
-      ? extraOverlayPath
-      : yield* _(
-        Effect.fail(
-          controllerBootstrapError(
-            `${controllerComposeExtraFileEnvKey} points to ${extraOverlayPath}, but it was not found.`
-          )
-        )
-      )
-  })
 
 const skillerSubmoduleCommand = [
   "submodule",
@@ -243,54 +207,6 @@ export const ensureSkillerSubmoduleInitialized = (
       )
     )
   })
-
-export const composeFilesForMode = (
-  composePath: string,
-  gpuOverlayPath: string | null,
-  runtimeOverlayPath: string | null = null,
-  extraOverlayPath: string | null = null
-): ReadonlyArray<string> => [
-  "-f",
-  composePath,
-  ...(runtimeOverlayPath === null ? [] : ["-f", runtimeOverlayPath]),
-  ...(gpuOverlayPath === null ? [] : ["-f", gpuOverlayPath]),
-  ...(extraOverlayPath === null ? [] : ["-f", extraOverlayPath])
-]
-
-export const composeFilesToArgs = (composeFiles: ControllerComposeFiles): ReadonlyArray<string> =>
-  composeFilesForMode(
-    composeFiles.composePath,
-    composeFiles.gpuOverlayPath,
-    composeFiles.runtimeOverlayPath,
-    composeFiles.extraOverlayPath
-  )
-
-const requireGpuOverlayPath = (
-  composePath: string
-): Effect.Effect<string, ControllerBootstrapError, FileSystem.FileSystem | Path.Path> =>
-  Effect.gen(function*(_) {
-    const fs = yield* _(FileSystem.FileSystem)
-    const path = yield* _(Path.Path)
-    const gpuOverlayPath = path.join(path.dirname(composePath), "docker-compose.gpu.yml")
-    const isExists = yield* _(fs.exists(gpuOverlayPath).pipe(Effect.mapError(mapComposePathError)))
-    return isExists
-      ? gpuOverlayPath
-      : yield* _(
-        Effect.fail(
-          controllerBootstrapError(`${controllerGpuModeEnvKey}=all requires ${gpuOverlayPath}, but it was not found.`)
-        )
-      )
-  })
-
-const composeFilesForGpuMode = (
-  composePath: string,
-  gpuMode: ControllerGpuMode
-): Effect.Effect<ControllerComposeFiles, ControllerBootstrapError, FileSystem.FileSystem | Path.Path> =>
-  gpuMode === "none"
-    ? Effect.succeed({ composePath, extraOverlayPath: null, gpuOverlayPath: null, runtimeOverlayPath: null })
-    : requireGpuOverlayPath(composePath).pipe(
-      Effect.map((gpuOverlayPath) => ({ composePath, extraOverlayPath: null, gpuOverlayPath, runtimeOverlayPath: null }))
-    )
 
 type ComposePathAndGpuMode = {
   readonly composePath: string
